refator: refatora componete Tab e Tabs
