+++ src/Components/Tabs/Tab.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from 'react';
+
+interface TabProps {
+  label: string;
+  children: ReactNode;
+}
+
+export const Tab: React.FC<TabProps> = ({ label, children }) => {
+  return (
+    <div data-label={label} className="hidden w-full">
+      {children}
+    </div>
+  );
+};
